requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==2.0.5
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.4.0
kombu==4.6.7
more-itertools==8.1.0
pkg-resources==0.0.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
urllib3==1.25.8
vine==1.3.0
zipp==1.0.0
